+++ app/components/data-menu/layer-directives/eventseries-directive.js
@@ -7,10 +7,21 @@ angular.module('data-menu')
 
     scope.remove = LayerAdderService.remove;
 
+    // Set defaults.
+    if (!scope.layer.name) {
+      scope.layer.name = scope.layer.type + ' ' + scope.layer.uuid;
+    }
+
     var cancelFirstActive = scope.$watch('layer.active', function () {
       if (scope.layer.active) {
         LayerAdderService.fetchLayer(scope.layer.type, scope.layer.uuid, scope.layer.name)
         .then(function (response) {
+
+          // If the layer did not have a name, check if the backend has one.
+          if (scope.layer.name === scope.layer.type + ' ' + scope.layer.uuid
+            && response.name) {
+            scope.layer.name = response.name;
+          }
 
           MapService.mapLayers.push(eventseriesMapLayer({
             color: response.color,
@@ -32,6 +43,9 @@ angular.module('data-menu')
             last: response.last_value_timestamp
           });
 
+        })
+        .catch(function () {
+          scope.invalid = true;
         });
 
         cancelFirstActive();