+++ app/components/data-menu/layer-directives/wmslayer-directive.js
@@ -18,12 +18,22 @@ angular.module('data-menu')
 
       scope.remove = LayerAdderService.remove;
 
+      // Set defaults.
       if (!scope.layer.opacity) { scope.layer.opacity = 1; }
+      if (!scope.layer.name) {
+        scope.layer.name = scope.layer.type + ' ' + scope.layer.uuid
+      }
 
       var cancelFirstActive = scope.$watch('layer.active', function () {
         if (scope.layer.active) {
           LayerAdderService.fetchLayer(scope.layer.type + 's', scope.layer.uuid, scope.layer.name)
           .then(function (response) {
+
+            // If the layer did not have a name, check if the backend has one.
+            if (scope.layer.name === scope.layer.type + ' ' + scope.layer.uuid
+              && response.name) {
+              scope.layer.name = response.name;
+            }
 
             MapService.mapLayers.push(wmsMapLayer({
               uuid: scope.layer.uuid,
@@ -36,6 +46,10 @@ angular.module('data-menu')
             }));
 
             MapService.updateLayers([scope.layer]);
+          })
+
+          .catch(function () {
+            scope.invalid = true;
           });
 
           cancelFirstActive();
@@ -43,7 +57,8 @@ angular.module('data-menu')
       });
 
       scope.$on('$destroy', function () {
-        // Remove layer from mapLayers and DataService
+        _.pull(DataService.dataLayers, {uuid: scope.layer.uuid });
+        _.pull(MapService.mapLayers, {uuid: scope.layer.uuid });
       });
 
     };