+++ app/components/map/map-directives.js
@@ -43,10 +43,10 @@ angular.module('map')
             maxZoom: 21,
             zoomControl: false,
             addZoomTitles: true,
+            worldCopyJump: true,
             lat: State.spatial.view.lat,
             lng: State.spatial.view.lng,
             zoom: State.spatial.view.zoom
-
           }, {
             onClick: _clicked,
             onMoveStart: _moveStarted,