+++ CHANGES.rst
@@ -1,9 +1,10 @@
 Changelog of lizard-nxt client
 ==============================
 
-Unreleased (2.9.0) (XXXX-XX-XX)
+Unreleased (4.0.2) (XXXX-XX-XX)
 -------------------------------
--
+
+- Fix: apply favourite when applying favourite, not state of favourite.
 
 
 Release 4.0.1 (2016-7-27)