+++ CHANGES.rst
@@ -12,6 +12,12 @@ Unreleased (4.0.2) (XXXX-XX-XX)
 
 - Fix not showing quantity in omnibox.
 
+- Fix adding layer which could not be fetched from api, catch and display error.
+
+- Include exclude parameter in layer-add request to filter out existing layers.
+
+- Include scenarios in layer adder.
+
 
 Release 4.0.1 (2016-7-27)
 ---------------------