+++ CHANGES.rst
@@ -1,9 +1,22 @@
 Changelog of lizard-nxt client
 ==============================
 
-Unreleased (2.9.0) (XXXX-XX-XX)
+Unreleased (4.0.2) (XXXX-XX-XX)
 -------------------------------
--
+
+- Fix: apply favourite when applying favourite, not state of favourite.
+
+- Fix request background tiles over https.
+
+- Fix panning around the world and messing up all coordinates.
+
+- Fix not showing quantity in omnibox.
+
+- Fix adding layer which could not be fetched from api, catch and display error.
+
+- Include exclude parameter in layer-add request to filter out existing layers.
+
+- Include scenarios in layer adder.
 
 
 Release 4.0.1 (2016-7-27)